# models/user.go
package models

import "time"

type User struct {
	Model
	Username  string `gorm:"column:username;not null;unique_index:username" json:"username" form:"username"`
	Firstname string `gorm:"column:firstname;not null;" json:"firstname" form:"firstname"`
	Lastname  string `gorm:"column:lastname;not null;" json:"lastname" form:"lastname"`
	Hash      string `gorm:"column:hash;not null;" json:"hash"`
	Role      string `gorm:"column:role;not null;" json:"role" form:"role"`
}

func (m *User) BeforeCreate() error {
	m.CreatedAt = time.Now()
	m.UpdatedAt = time.Now()
	return nil
}

func (m *User) BeforeUpdate() error {
	m.UpdatedAt = time.Now()
	return nil
}